Banner/Modal Feature Static Data - Complete
Implemented both a banner and modal that will appear for users in UserGroups= 0001-0008 or 0001-0005 (PROD). The message and user groups are currently hardcoded into the solution. The goal will be to implement dynamic data functionality so that Admin Users will be able to change the promo banner/modal message via UI in the Admin app as well as assign which user Groups will be able to view the banner/modal.